requirements.txt
requests
ratelimit
tqdm
pathos
secedgar
pandas
html2text
readability-lxml
bs4